refactor(travel): /simplify on Phase 8 — kill regex, tighten path check
Three findings from the parallel /simplify pass acted on.

1. Replaced regex-based VEVENT extraction with a proper helper.
   TripSummary's Export Calendar was calling buildIcsContent() per event
   and then regex-stripping the VCALENDAR envelope with
   `/BEGIN:VEVENT[\s\S]*?END:VEVENT/` — brittle (depends on an internal
   shape of buildIcsContent staying stable) and unnecessary.

   Refactored src/lib/calendar.ts to expose buildVeventBlock(event) as
   the primitive; buildIcsContent wraps one VEVENT in a VCALENDAR
   envelope, and a new exported buildMultiEventIcs(events) wraps N
   VEVENTs under one envelope. TripSummary calls buildMultiEventIcs
   directly — no regex, no dependency on internal string shapes, fully
   unit-testable from calendar.ts's existing test coverage.

2. Tightened parseTravelRedirect's path check.
   `url.pathname.startsWith("/travel")` was loose — matched /travellers,
   /travel-guide, /travelogue etc. In practice the required q/from/to
   params guarded against false positives, but the intent was unclear.
   Now: `pathname === "/travel" || pathname.startsWith("/travel/")`.

3. Moved parseTravelRedirect to src/lib/travel/url.ts so the
   URL builder and parser live in one module. Colocating
   buildTravelSearchUrl + parseTravelRedirect makes the param contract
   with /travel/page.tsx explicit — both sides read from the same file.
   Removed the stale page.test.ts; consolidated tests into
   url.test.ts with a new regression case for the /travellers false-
   match.

Also improved the comment on TravelResults' results_viewed effect to
explain the ref-guard + viewedKey pattern: the string key prevents
re-fire on filter toggles that don't change counts; the ref guard is
defense-in-depth for React Strict Mode double-mount. Deps stay
exhaustive-compliant rather than `[viewedKey]` — not worth an
eslint-disable.

Skipped (per simplify findings):
- slugifyForFilename extraction — one caller, not worth a shared module
- Filter action closure consolidation — three 2-3 line closures, inline
  reads cleaner than a shared helper
- auto-save + delete pattern extraction — diverges on router.replace
  vs router.refresh
- Banner `isSave` boolean → string union — fine with two modes
