+++ src/lib/travel/url.test.ts
@@ -1,5 +1,9 @@
 import { describe, it, expect, vi } from "vitest";
-import { buildTravelSearchUrl, withConcurrency } from "./url";
+import {
+  buildTravelSearchUrl,
+  parseTravelRedirect,
+  withConcurrency,
+} from "./url";
 
 describe("buildTravelSearchUrl", () => {
   it("emits the param names /travel page.tsx parses", () => {
@@ -93,5 +97,72 @@ describe("withConcurrency", () => {
     const out = await withConcurrency([], 5, fn);
     expect(out).toEqual([]);
     expect(fn).not.toHaveBeenCalled();
+  });
+});
+
+describe("parseTravelRedirect", () => {
+  it("returns null for a null redirect", () => {
+    expect(parseTravelRedirect(null)).toBeNull();
+  });
+
+  it("returns null when the path doesn't target /travel", () => {
+    expect(
+      parseTravelRedirect("/hareline?q=Boston&from=2026-04-12&to=2026-04-20"),
+    ).toBeNull();
+  });
+
+  it("rejects paths that merely start with 'travel' (e.g. /travellers)", () => {
+    // Regression: earlier impl used `startsWith("/travel")` which matched
+    // /travellers, /travel-something, etc. Now uses exact match or
+    // /travel/ prefix so those false positives are blocked.
+    expect(
+      parseTravelRedirect(
+        "/travellers?q=Boston&from=2026-04-12&to=2026-04-20",
+      ),
+    ).toBeNull();
+    expect(
+      parseTravelRedirect(
+        "/travel-guide?q=Boston&from=2026-04-12&to=2026-04-20",
+      ),
+    ).toBeNull();
+  });
+
+  it("returns null when required params are missing", () => {
+    expect(parseTravelRedirect("/travel?from=2026-04-12&to=2026-04-20")).toBeNull();
+    expect(parseTravelRedirect("/travel?q=Boston&to=2026-04-20")).toBeNull();
+    expect(parseTravelRedirect("/travel?q=Boston&from=2026-04-12")).toBeNull();
+  });
+
+  it("extracts destination + dates when /travel params are present", () => {
+    expect(
+      parseTravelRedirect(
+        "/travel?lat=42.3&lng=-71.0&q=Boston%2C+MA%2C+USA&from=2026-04-12&to=2026-04-20",
+      ),
+    ).toEqual({
+      destination: "Boston, MA, USA",
+      startDate: "2026-04-12",
+      endDate: "2026-04-20",
+      isSave: false,
+    });
+  });
+
+  it("flags isSave when saved=1 is present", () => {
+    expect(
+      parseTravelRedirect(
+        "/travel?q=Boston&from=2026-04-12&to=2026-04-20&saved=1",
+      )?.isSave,
+    ).toBe(true);
+  });
+
+  it("accepts /travel/saved paths (exact-or-/travel/-prefix)", () => {
+    const out = parseTravelRedirect(
+      "/travel/saved?q=Boston&from=2026-04-12&to=2026-04-20",
+    );
+    expect(out).not.toBeNull();
+    expect(out?.destination).toBe("Boston");
+  });
+
+  it("returns null for malformed URLs", () => {
+    expect(parseTravelRedirect(":not a url")).toBeNull();
   });
 });
